feat(cire/web): design registry + classic pack; [slug] renders the resolved design
