+++ cire/web/src/designs/classic/UnlockReveal.motion.test.ts
