+++ cire/web/src/designs/classic/UnlockReveal.motion.ts
